Added: - auto-paste feature in cisco any connect for faster VPN connection; - lock-button for set/remove secret - auto-copy code button - notification icon in system tray bar - foreground-working feature for easy-accessing to app anytime Release v1.0

+++ OTPGenerator/MainWindow.xaml.cs
@@ -1,4 +1,7 @@
-﻿using System.Windows;
+﻿using System.ComponentModel;
+using System.Media;
+using System.Windows;
+using System.Windows.Media;
 using System.Windows.Threading;
 
 namespace OTPGenerator;
@@ -6,9 +9,17 @@ namespace OTPGenerator;
 public partial class MainWindow : Window
 {
     private const string CredentialName = "OTPGeneratorSecretKey";
+    private const string SecretMask = "************";
     private string? _secretKey;
     private readonly DispatcherTimer _timer;
     private readonly OtpManager _otpManager;
+
+    protected override void OnClosing(CancelEventArgs e)
+    {
+        e.Cancel = true;
+        Hide();
+        base.OnClosing(e);
+    }
 
     public MainWindow()
     {
@@ -22,6 +33,8 @@ public partial class MainWindow : Window
         _timer.Start();
 
         LoadSecretKey();
+        SetSecretButton.IsEnabled = false;
+        ClearSecretButton.IsEnabled = false;
     }
 
     private void LoadSecretKey()
@@ -29,13 +42,13 @@ public partial class MainWindow : Window
         _secretKey = Credential.RetrieveCredential(CredentialName);
         if (!string.IsNullOrEmpty(_secretKey))
         {
-            SecretKeyLabel.Text = "********"; // Mask the secret key
+            SecretKeyLabel.Text = SecretMask; // Mask the secret key
             UpdateOtpCode();
             UpdateStatus("Secret key loaded successfully.");
         }
         else
         {
-            UpdateStatus("No secret key found. Please set a new one.");
+            UpdateStatus("No secret key found. Please set a new one.", false, true);
         }
     }
 
@@ -46,6 +59,7 @@ public partial class MainWindow : Window
         {
             _secretKey = key;
             UpdateStatus("Secret key saved successfully.");
+            SecretKeyLabel.Text = SecretMask;
         }
         else
         {
@@ -66,14 +80,14 @@ public partial class MainWindow : Window
         UpdateOtpCode();
     }
 
-    private void UpdateOtpCode()
+    private void UpdateOtpCode(bool forceRefresh = false)
     {
         if (string.IsNullOrWhiteSpace(_secretKey))
         {
             return;
         }
 
-        var otpCode = _otpManager.GenerateOtp(_secretKey);
+        var otpCode = _otpManager.GenerateOtp(_secretKey, forceRefresh);
         OtpCodeLabel.Text = otpCode;
         var timeLeft = 30 - (int)(DateTimeOffset.UtcNow.ToUnixTimeSeconds() % 30);
         TimeoutBar.Value = timeLeft;
@@ -82,6 +96,12 @@ public partial class MainWindow : Window
 
     private void OnClearSecretButtonClick(object sender, RoutedEventArgs e)
     {
+        if (MessageBox.Show("Are you sure you want to clear the secret key?", "Confirmation", MessageBoxButton.YesNo)
+            != MessageBoxResult.OK)
+        {
+            return;
+        }
+
         if (Credential.DeleteCredential(CredentialName))
         {
             _secretKey = null;
@@ -93,7 +113,7 @@ public partial class MainWindow : Window
         }
         else
         {
-            UpdateStatus("Failed to clear secret key.");
+            UpdateStatus("Failed to clear secret key.", false, true);
         }
     }
 
@@ -108,9 +128,78 @@ public partial class MainWindow : Window
     }
 
 
-    private void UpdateStatus(string message)
-        => StatusLabel.Text = message;
+    private void UpdateStatus(string message, bool calmMessage = true, bool playExclamationSound = false)
+    {
+        StatusLabel.Text = message;
+        StatusLabel.Foreground = calmMessage
+            ? Brushes.Green
+            : Brushes.Red;
+
+        if (playExclamationSound)
+        {
+            SystemSounds.Exclamation.Play();
+        }
+    }
+
+    private void OnEnableAutoPastingInCiscoRadioButtonClick(object sender, RoutedEventArgs e)
+        => EnableAutoPastingInCiscoRadioButton.IsChecked = !EnableAutoPastingInCiscoRadioButton.IsChecked;
 
     private void OnCopyToClipboardButtonClick(object sender, RoutedEventArgs e)
-        => Clipboard.SetText(OtpCodeLabel.Text);
+    {
+        Clipboard.SetText(OtpCodeLabel.Text);
+        if (!EnableAutoPastingInCiscoRadioButton.IsChecked.HasValue ||
+            !EnableAutoPastingInCiscoRadioButton.IsChecked.Value)
+        {
+            UpdateStatus("The verification code has been copied to the clipboard.");
+            return;
+        }
+
+        var hWnd = WindowUtils.FindWindow(@"Cisco AnyConnect \| TAXSEE\d+");
+        if (hWnd != IntPtr.Zero)
+        {
+            var hEdit = FindThirdEditControl(hWnd);
+            if (hEdit != IntPtr.Zero)
+            {
+                WindowUtils.SetForegroundWindow(hWnd);
+                WindowUtils.SendMessageNative(hEdit, 0x000C, 0, OtpCodeLabel.Text);
+                UpdateStatus("The verification code has been inserted successfully!");
+            }
+            else
+            {
+                UpdateStatus("The `Second Password` input field could not be found.", false, true);
+            }
+        }
+        else
+        {
+            UpdateStatus("The Cisco AnyConnect code entry window was not found.", false, true);
+        }
+    }
+
+    private void OnUnlockButtonClick(object sender, RoutedEventArgs e)
+    {
+        SetSecretButton.IsEnabled = !SetSecretButton.IsEnabled;
+        ClearSecretButton.IsEnabled = !ClearSecretButton.IsEnabled;
+    }
+
+    private static IntPtr FindThirdEditControl(IntPtr hWnd)
+    {
+        IntPtr hEdit;
+
+        var hStatic = WindowUtils.FindWindowExNative(hWnd, IntPtr.Zero, "Static", "Second Password:");
+
+        if (hStatic != IntPtr.Zero)
+        {
+            hEdit = WindowUtils.FindWindowExNative(hWnd, hStatic, "Edit", null);
+        }
+        else
+        {
+            hEdit = WindowUtils.FindWindowExNative(hWnd, IntPtr.Zero, "Edit", null);
+            for (var i = 0; i < 2; i++)
+            {
+                hEdit = WindowUtils.FindWindowExNative(hWnd, hEdit, "Edit", null);
+            }
+        }
+
+        return hEdit;
+    }
 }
